+++ app/src/androidTest/java/org/kiwix/kiwixmobile/main/KiwixMainActivityTest.kt
@@ -55,7 +55,7 @@ class KiwixMainActivityTest {
       openDrawer()
       clickSettingsOnSideNav { }
       // assertSettingsScreenDisplayed()
-      navigate()
+      openDrawer()
       clickHelpOnSideNav()
       assertHelpScreenDisplayed()
       navigate()
